sia20: improve mesh quality.
This change adds:

 - detailed meshes based on CAD model offered on motoman.co.uk web site
 - updated collision meshes based on the detailed meshes

Also minor fixes to the way materials / mesh colours are configured for
the links in the collision model (must be set on the 'visual' geometry
elements, but colours are only applied to the 'collision' geometry).

No changes to the kinematic chain itself.

Thanks to @ted-miller from Motoman for providing an updated CAD model
of the SIA20.